+++ lizard_fewsjdbc/migrations/0007_auto__add_timeseriescache__add_webrssource__add_parametercache__add_lo.py
@@ -14,51 +14,56 @@ class Migration(SchemaMigration):
             ('t_filter', self.gf('django.db.models.fields.related.ForeignKey')(to=orm['lizard_fewsjdbc.FilterCache'])),
             ('t_location', self.gf('django.db.models.fields.related.ForeignKey')(to=orm['lizard_fewsjdbc.LocationCache'])),
             ('t_parameter', self.gf('django.db.models.fields.related.ForeignKey')(to=orm['lizard_fewsjdbc.ParameterCache'])),
+            ('webrs_source', self.gf('django.db.models.fields.related.ForeignKey')(to=orm['lizard_fewsjdbc.WebRSSource'], null=True, blank=True)),
         ))
         db.send_create_signal('lizard_fewsjdbc', ['TimeseriesCache'])
 
         # Adding model 'WebRSSource'
         db.create_table('lizard_fewsjdbc_webrssource', (
-            ('code', self.gf('django.db.models.fields.CharField')(max_length=50, primary_key=True)),
+            ('id', self.gf('django.db.models.fields.AutoField')(primary_key=True)),
             ('name', self.gf('django.db.models.fields.CharField')(max_length=50)),
+            ('slug', self.gf('django.db.models.fields.SlugField')(max_length=50)),
             ('version', self.gf('django.db.models.fields.CharField')(max_length=10)),
-            ('base_path', self.gf('django.db.models.fields.URLField')(max_length=250)),
+            ('base_path', self.gf('django.db.models.fields.CharField')(max_length=250)),
         ))
         db.send_create_signal('lizard_fewsjdbc', ['WebRSSource'])
 
         # Adding model 'ParameterCache'
         db.create_table('lizard_fewsjdbc_parametercache', (
             ('parameterid', self.gf('django.db.models.fields.CharField')(max_length=50, primary_key=True)),
-            ('name', self.gf('django.db.models.fields.CharField')(max_length=50, null=True, blank=True)),
-            ('short_name', self.gf('django.db.models.fields.CharField')(max_length=50, null=True, blank=True)),
-            ('unit', self.gf('django.db.models.fields.CharField')(max_length=50, null=True, blank=True)),
-            ('parameter_type', self.gf('django.db.models.fields.CharField')(max_length=50, null=True, blank=True)),
-            ('parameter_group', self.gf('django.db.models.fields.CharField')(max_length=50, null=True, blank=True)),
+            ('name', self.gf('django.db.models.fields.CharField')(max_length=100, null=True, blank=True)),
+            ('short_name', self.gf('django.db.models.fields.CharField')(max_length=100, null=True, blank=True)),
+            ('unit', self.gf('django.db.models.fields.CharField')(max_length=100, null=True, blank=True)),
+            ('parameter_type', self.gf('django.db.models.fields.CharField')(max_length=100, null=True, blank=True)),
+            ('parameter_group', self.gf('django.db.models.fields.CharField')(max_length=100, null=True, blank=True)),
+            ('webrs_source', self.gf('django.db.models.fields.related.ForeignKey')(to=orm['lizard_fewsjdbc.WebRSSource'], null=True, blank=True)),
         ))
         db.send_create_signal('lizard_fewsjdbc', ['ParameterCache'])
 
         # Adding model 'LocationCache'
         db.create_table('lizard_fewsjdbc_locationcache', (
-            ('locationid', self.gf('django.db.models.fields.CharField')(max_length=50, primary_key=True)),
-            ('name', self.gf('django.db.models.fields.CharField')(max_length=50, null=True, blank=True)),
-            ('short_name', self.gf('django.db.models.fields.CharField')(max_length=50, null=True, blank=True)),
+            ('locationid', self.gf('django.db.models.fields.CharField')(max_length=100, primary_key=True)),
+            ('name', self.gf('django.db.models.fields.CharField')(max_length=100, null=True, blank=True)),
+            ('short_name', self.gf('django.db.models.fields.CharField')(max_length=100, null=True, blank=True)),
             ('description', self.gf('django.db.models.fields.CharField')(max_length=250, null=True, blank=True)),
             ('lng', self.gf('django.db.models.fields.FloatField')(null=True, blank=True)),
             ('lat', self.gf('django.db.models.fields.FloatField')(null=True, blank=True)),
             ('tooltiptext', self.gf('django.db.models.fields.TextField')(null=True, blank=True)),
-            ('parent_id', self.gf('django.db.models.fields.CharField')(max_length=50, null=True, blank=True)),
+            ('parent_id', self.gf('django.db.models.fields.CharField')(max_length=100, null=True, blank=True)),
+            ('webrs_source', self.gf('django.db.models.fields.related.ForeignKey')(to=orm['lizard_fewsjdbc.WebRSSource'], null=True, blank=True)),
         ))
         db.send_create_signal('lizard_fewsjdbc', ['LocationCache'])
 
         # Adding model 'FilterCache'
         db.create_table('lizard_fewsjdbc_filtercache', (
-            ('filterid', self.gf('django.db.models.fields.CharField')(max_length=50, primary_key=True)),
-            ('name', self.gf('django.db.models.fields.CharField')(max_length=50, null=True, blank=True)),
-            ('description', self.gf('django.db.models.fields.CharField')(max_length=50, null=True, blank=True)),
+            ('filterid', self.gf('django.db.models.fields.CharField')(max_length=100, primary_key=True)),
+            ('name', self.gf('django.db.models.fields.CharField')(max_length=100, null=True, blank=True)),
+            ('description', self.gf('django.db.models.fields.CharField')(max_length=100, null=True, blank=True)),
             ('is_sub_filter', self.gf('django.db.models.fields.BooleanField')(default=False)),
-            ('parent_id', self.gf('django.db.models.fields.CharField')(max_length=50, null=True, blank=True)),
-            ('parent_name', self.gf('django.db.models.fields.CharField')(max_length=50, null=True, blank=True)),
+            ('parent_id', self.gf('django.db.models.fields.CharField')(max_length=100, null=True, blank=True)),
+            ('parent_name', self.gf('django.db.models.fields.CharField')(max_length=100, null=True, blank=True)),
             ('is_end_node', self.gf('django.db.models.fields.BooleanField')(default=False)),
+            ('webrs_source', self.gf('django.db.models.fields.related.ForeignKey')(to=orm['lizard_fewsjdbc.WebRSSource'], null=True, blank=True)),
         ))
         db.send_create_signal('lizard_fewsjdbc', ['FilterCache'])
 
@@ -83,13 +88,14 @@ class Migration(SchemaMigration):
     models = {
         'lizard_fewsjdbc.filtercache': {
             'Meta': {'object_name': 'FilterCache'},
-            'description': ('django.db.models.fields.CharField', [], {'max_length': '50', 'null': 'True', 'blank': 'True'}),
-            'filterid': ('django.db.models.fields.CharField', [], {'max_length': '50', 'primary_key': 'True'}),
+            'description': ('django.db.models.fields.CharField', [], {'max_length': '100', 'null': 'True', 'blank': 'True'}),
+            'filterid': ('django.db.models.fields.CharField', [], {'max_length': '100', 'primary_key': 'True'}),
             'is_end_node': ('django.db.models.fields.BooleanField', [], {'default': 'False'}),
             'is_sub_filter': ('django.db.models.fields.BooleanField', [], {'default': 'False'}),
-            'name': ('django.db.models.fields.CharField', [], {'max_length': '50', 'null': 'True', 'blank': 'True'}),
-            'parent_id': ('django.db.models.fields.CharField', [], {'max_length': '50', 'null': 'True', 'blank': 'True'}),
-            'parent_name': ('django.db.models.fields.CharField', [], {'max_length': '50', 'null': 'True', 'blank': 'True'})
+            'name': ('django.db.models.fields.CharField', [], {'max_length': '100', 'null': 'True', 'blank': 'True'}),
+            'parent_id': ('django.db.models.fields.CharField', [], {'max_length': '100', 'null': 'True', 'blank': 'True'}),
+            'parent_name': ('django.db.models.fields.CharField', [], {'max_length': '100', 'null': 'True', 'blank': 'True'}),
+            'webrs_source': ('django.db.models.fields.related.ForeignKey', [], {'to': "orm['lizard_fewsjdbc.WebRSSource']", 'null': 'True', 'blank': 'True'})
         },
         'lizard_fewsjdbc.iconstyle': {
             'Meta': {'object_name': 'IconStyle'},
@@ -120,20 +126,22 @@ class Migration(SchemaMigration):
             'description': ('django.db.models.fields.CharField', [], {'max_length': '250', 'null': 'True', 'blank': 'True'}),
             'lat': ('django.db.models.fields.FloatField', [], {'null': 'True', 'blank': 'True'}),
             'lng': ('django.db.models.fields.FloatField', [], {'null': 'True', 'blank': 'True'}),
-            'locationid': ('django.db.models.fields.CharField', [], {'max_length': '50', 'primary_key': 'True'}),
-            'name': ('django.db.models.fields.CharField', [], {'max_length': '50', 'null': 'True', 'blank': 'True'}),
-            'parent_id': ('django.db.models.fields.CharField', [], {'max_length': '50', 'null': 'True', 'blank': 'True'}),
-            'short_name': ('django.db.models.fields.CharField', [], {'max_length': '50', 'null': 'True', 'blank': 'True'}),
-            'tooltiptext': ('django.db.models.fields.TextField', [], {'null': 'True', 'blank': 'True'})
+            'locationid': ('django.db.models.fields.CharField', [], {'max_length': '100', 'primary_key': 'True'}),
+            'name': ('django.db.models.fields.CharField', [], {'max_length': '100', 'null': 'True', 'blank': 'True'}),
+            'parent_id': ('django.db.models.fields.CharField', [], {'max_length': '100', 'null': 'True', 'blank': 'True'}),
+            'short_name': ('django.db.models.fields.CharField', [], {'max_length': '100', 'null': 'True', 'blank': 'True'}),
+            'tooltiptext': ('django.db.models.fields.TextField', [], {'null': 'True', 'blank': 'True'}),
+            'webrs_source': ('django.db.models.fields.related.ForeignKey', [], {'to': "orm['lizard_fewsjdbc.WebRSSource']", 'null': 'True', 'blank': 'True'})
         },
         'lizard_fewsjdbc.parametercache': {
             'Meta': {'object_name': 'ParameterCache'},
-            'name': ('django.db.models.fields.CharField', [], {'max_length': '50', 'null': 'True', 'blank': 'True'}),
-            'parameter_group': ('django.db.models.fields.CharField', [], {'max_length': '50', 'null': 'True', 'blank': 'True'}),
-            'parameter_type': ('django.db.models.fields.CharField', [], {'max_length': '50', 'null': 'True', 'blank': 'True'}),
+            'name': ('django.db.models.fields.CharField', [], {'max_length': '100', 'null': 'True', 'blank': 'True'}),
+            'parameter_group': ('django.db.models.fields.CharField', [], {'max_length': '100', 'null': 'True', 'blank': 'True'}),
+            'parameter_type': ('django.db.models.fields.CharField', [], {'max_length': '100', 'null': 'True', 'blank': 'True'}),
             'parameterid': ('django.db.models.fields.CharField', [], {'max_length': '50', 'primary_key': 'True'}),
-            'short_name': ('django.db.models.fields.CharField', [], {'max_length': '50', 'null': 'True', 'blank': 'True'}),
-            'unit': ('django.db.models.fields.CharField', [], {'max_length': '50', 'null': 'True', 'blank': 'True'})
+            'short_name': ('django.db.models.fields.CharField', [], {'max_length': '100', 'null': 'True', 'blank': 'True'}),
+            'unit': ('django.db.models.fields.CharField', [], {'max_length': '100', 'null': 'True', 'blank': 'True'}),
+            'webrs_source': ('django.db.models.fields.related.ForeignKey', [], {'to': "orm['lizard_fewsjdbc.WebRSSource']", 'null': 'True', 'blank': 'True'})
         },
         'lizard_fewsjdbc.threshold': {
             'Meta': {'ordering': "('id',)", 'object_name': 'Threshold'},
@@ -151,13 +159,15 @@ class Migration(SchemaMigration):
             'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
             't_filter': ('django.db.models.fields.related.ForeignKey', [], {'to': "orm['lizard_fewsjdbc.FilterCache']"}),
             't_location': ('django.db.models.fields.related.ForeignKey', [], {'to': "orm['lizard_fewsjdbc.LocationCache']"}),
-            't_parameter': ('django.db.models.fields.related.ForeignKey', [], {'to': "orm['lizard_fewsjdbc.ParameterCache']"})
+            't_parameter': ('django.db.models.fields.related.ForeignKey', [], {'to': "orm['lizard_fewsjdbc.ParameterCache']"}),
+            'webrs_source': ('django.db.models.fields.related.ForeignKey', [], {'to': "orm['lizard_fewsjdbc.WebRSSource']", 'null': 'True', 'blank': 'True'})
         },
         'lizard_fewsjdbc.webrssource': {
             'Meta': {'object_name': 'WebRSSource'},
-            'base_path': ('django.db.models.fields.URLField', [], {'max_length': '250'}),
-            'code': ('django.db.models.fields.CharField', [], {'max_length': '50', 'primary_key': 'True'}),
+            'base_path': ('django.db.models.fields.CharField', [], {'max_length': '250'}),
+            'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
             'name': ('django.db.models.fields.CharField', [], {'max_length': '50'}),
+            'slug': ('django.db.models.fields.SlugField', [], {'max_length': '50'}),
             'version': ('django.db.models.fields.CharField', [], {'max_length': '10'})
         }
     }